query 'overwrite?' if specified directory exists when creating new project

+++ core/fileutils.cpp
@@ -6,10 +6,11 @@
 
 /**
  * remove a directory along with all its contents
- * @param   dir path of directory to remove
+ * @param   dir     path of directory to remove
+ * @param   rmSelf  remove dir and its contents if true; remove contents only and keep dir itself if false
  * @return  true on success; false on errors
  */
-bool FileUtils::removeDir(const QString &dir)
+bool FileUtils::removeDir(const QString &dir, bool rmSelf)
 {
     bool result = true;
     QDir path(dir);
@@ -21,13 +22,14 @@ bool FileUtils::removeDir(const QString &dir)
                                      QDir::DirsFirst))
         {
             if (info.isDir())
-                result = removeDir(info.absoluteFilePath());
+                result = removeDir(info.absoluteFilePath(), true);
             else
                 result = QFile::remove(info.absoluteFilePath());
             if (!result)
                 return result;
         }
-        result = path.rmdir(dir);
+        if (rmSelf)
+            result = path.rmdir(dir);
     }
     return result;
 }
